2022-11-24_10-42PM@chan-maddanna:
'fix' Branch commit details

The vulnerability I choose to fix is plaintext credentials, as a ConfigMap data
present in the file k8s/base-config.yaml. In a situation with multiple microservices
and a dedicated IAM/SSO in the cluster (example SAML2/oAuth2 etc), I would have
preferred to user ExternalSecrets objects with an ExternalSecretProvider.
(example: Hashicorp Vault with point-in-time seal/unseal )

However, in this case to showcase an example, of how to seal and secure secrets
even in environments where they might be heavily hardcoded and vulnerable, without
breaking or disturbing any other component, I have used SealedSecrets with a
SealedSecretController running on the cluster, which is equally cryptographically
secure and more suitable for the use case.

Infact, I have used sealed secrets in more then 2 SMB's in Canada, among a host of
other measures, to achieve ISO/IEC 21000.

FEW OTHER FIXES (that i could not resist)

On the way I also happenned to clean up a few stuff:
- Updated Dockerfile to run as non root user.
- Updated Dockerfile to use an updated base image that
  has many vulnerabilities fixed.
- Updated Dockerfile to cleanup unnecessary packages
  that have known HIGH/CRITICAL cve's.
- Updated Makefile to follow variable and naming standards.
- Updated Makefile to read from version.txt for version setting.
- Updated Makefile to include Basic SAST, Linting and Changelog
- Added checks for data race condition check.
- Added internal target for to-be allocated Memory Sanitization.